featTmpHome 정리
